+++ .drone.yml
@@ -1,15 +1,15 @@
 image: bjodah/dockerfile-bjodah-scicomp:14.04c
 git:
-    path: github.com/bjodah/aqchem
+    path: github.com/bjodah/chempy
 script:
-  - if [[ "$DRONE_BRANCH" =~ ^v[0-9]+.[0-9]?* ]]; then export AQCHEM_RELEASE_VERSION=$DRONE_BRANCH; echo ${AQCHEM_RELEASE_VERSION} | tail -c +2 > __conda_version__.txt; fi
+  - if [[ "$DRONE_BRANCH" =~ ^v[0-9]+.[0-9]?* ]]; then export CHEMPY_RELEASE_VERSION=$DRONE_BRANCH; echo ${CHEMPY_RELEASE_VERSION} | tail -c +2 > __conda_version__.txt; fi
   - pip install --user -r requirements.txt
   - pip3 install --user -r requirements.txt
-  - PYTHONPATH=$(pwd) ./scripts/run_tests.sh --cov aqchem --cov-report html
+  - PYTHONPATH=$(pwd) ./scripts/run_tests.sh --cov chempy --cov-report html
   - python2 setup.py sdist
   # - ./scripts/generate_docs.sh
-  # - scp -r doc/_build/html/ dist/* htmlcov/ aqchem@hera.physchem.kth.se:~/public_html/
+  # - scp -r doc/_build/html/ dist/* htmlcov/ chempy@hera.physchem.kth.se:~/public_html/
   # - touch doc/_build/html/.nojekyll
   # - git config --global user.name "drone"
   # - git config --global user.email "drone@hera.physchem.kth.se"
-  # - if [[ "$DRONE_BRANCH" == "master" ]]; then ./scripts/github_upload.sh doc/_build/html bjodah aqchem gh-pages; fi
+  # - if [[ "$DRONE_BRANCH" == "master" ]]; then ./scripts/github_upload.sh doc/_build/html bjodah chempy gh-pages; fi
