--- aqchem/__init__.py
@@ -1,8 +1,0 @@
-"""
-aqchem is a Python package useful for (physical) chemistry.
-"""
-
-from ._release import __version__
-__url__ = "https://github.com/bjodah/aqchem"
-
-assert __version__  # silence pyflakes
