+++ chempy/__init__.py
@@ -1,0 +1,8 @@
+"""
+chempy is a Python package useful for (physical) chemistry.
+"""
+
+from ._release import __version__
+__url__ = "https://github.com/bjodah/chempy"
+
+assert __version__  # silence pyflakes
