+++ chempy/kinetics/integrated.py
