+++ chempy/kinetics/tests/test_integrated.py
