+++ chempy/tests/test_core.py
