+++ chempy/tests/test_equilibria.py
