+++ chempy/tests/util.py
