+++ examples/demo_kinetics.py
@@ -4,7 +4,7 @@
 from __future__ import (absolute_import, division,
                         print_function, unicode_literals)
 
-from aqchem.kinetics import pseudo_irrev, pseudo_rev, binary_irrev, binary_rev
+from chempy.kinetics import pseudo_irrev, pseudo_rev, binary_irrev, binary_rev
 
 import sympy
